real-estate-visualizer/backend/app/api/render.py
from fastapi import APIRouter, HTTPException, UploadFile, File
from fastapi.responses import JSONResponse
from pydantic import BaseModel
from typing import List, Optional
import base64
import io
from PIL import Image
import numpy as np

router = APIRouter(prefix="/api/render", tags=["rendering"])

class RenderRequest(BaseModel):
    property_type: str
    style: str
    room_layout: dict
    materials: dict
    lighting: dict
    resolution: str = "1024x1024"

class RenderResponse(BaseModel):
    image_b64: str
    render_time: float
    metadata: dict

@router.post("/2d", response_model=RenderResponse)
async def render_2d_property(request: RenderRequest):
    """
    Generate 2D property visualization using AI
    """
    try:
        # TODO: Integrate with AI 2D renderer
        # For now, return a placeholder
        placeholder_image = create_placeholder_image(request.resolution)
        
        return RenderResponse(
            image_b64=placeholder_image,
            render_time=1.5,
            metadata={
                "property_type": request.property_type,
                "style": request.style,
                "resolution": request.resolution
            }
        )
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.post("/3d", response_model=RenderResponse)
async def render_3d_property(request: RenderRequest):
    """
    Generate 3D property visualization using AI
    """
    try:
        # TODO: Integrate with AI 3D renderer
        placeholder_image = create_placeholder_image(request.resolution)
        
        return RenderResponse(
            image_b64=placeholder_image,
            render_time=3.2,
            metadata={
                "property_type": request.property_type,
                "style": request.style,
                "resolution": request.resolution,
                "is_3d": True
            }
        )
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@router.post("/upload")
async def upload_property_image(file: UploadFile = File(...)):
    """
    Upload property image for AI analysis and enhancement
    """
    try:
        # Validate file type
        if not file.content_type.startswith('image/'):
            raise HTTPException(status_code=400, detail="File must be an image")
        
        # Read and process image
        contents = await file.read()
        image = Image.open(io.BytesIO(contents))
        
        # TODO: Process with AI for enhancement
        enhanced_image_b64 = create_placeholder_image("1024x1024")
        
        return JSONResponse({
            "message": "Image uploaded successfully",
            "enhanced_image": enhanced_image_b64,
            "original_size": image.size,
            "file_name": file.filename
        })
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

def create_placeholder_image(resolution: str) -> str:
    """Create a placeholder image for testing"""
    width, height = map(int, resolution.split('x'))
    
    # Create a simple gradient image
    image = Image.new('RGB', (width, height), color='lightblue')
    
    # Convert to base64
    buffer = io.BytesIO()
    image.save(buffer, format='PNG')
    img_str = base64.b64encode(buffer.getvalue()).decode()
    
    return img_str
