real-estate-visualizer/backend/app/services/ai_engine.py
"""
AI Engine Service for Real Estate Visualization
Handles 2D/3D rendering, image enhancement, and property analysis
"""

import asyncio
import base64
import io
from typing import Dict, List, Optional, Tuple
from PIL import Image, ImageDraw, ImageFont
import numpy as np
import cv2

class AIEngine:
    def __init__(self):
        self.model_loaded = False
        # TODO: Initialize actual AI models here
        
    async def initialize_models(self):
        """Initialize AI models for rendering and analysis"""
        try:
            # TODO: Load actual AI models
            # - 2D property renderer
            # - 3D property generator
            # - Image enhancement model
            # - Property analysis model
            await asyncio.sleep(1)  # Simulate model loading
            self.model_loaded = True
            return True
        except Exception as e:
            print(f"Error initializing AI models: {e}")
            return False
    
    async def render_2d_property(
        self, 
        property_data: Dict, 
        style: str = "modern",
        resolution: Tuple[int, int] = (1024, 1024)
    ) -> str:
        """
        Generate 2D property visualization
        """
        try:
            if not self.model_loaded:
                await self.initialize_models()
            
            # TODO: Use actual AI model for 2D rendering
            # For now, create a sophisticated placeholder
            image = await self._create_2d_placeholder(property_data, style, resolution)
            
            # Convert to base64
            buffer = io.BytesIO()
            image.save(buffer, format='PNG')
            img_str = base64.b64encode(buffer.getvalue()).decode()
            
            return img_str
            
        except Exception as e:
            raise Exception(f"2D rendering failed: {str(e)}")
    
    async def render_3d_property(
        self, 
        property_data: Dict, 
        style: str = "modern",
        resolution: Tuple[int, int] = (1024, 1024)
    ) -> str:
        """
        Generate 3D property visualization
        """
        try:
            if not self.model_loaded:
                await self.initialize_models()
            
            # TODO: Use actual AI model for 3D rendering
            # For now, create a 3D-style placeholder
            image = await self._create_3d_placeholder(property_data, style, resolution)
            
            # Convert to base64
            buffer = io.BytesIO()
            image.save(buffer, format='PNG')
            img_str = base64.b64encode(buffer.getvalue()).decode()
            
            return img_str
            
        except Exception as e:
            raise Exception(f"3D rendering failed: {str(e)}")
    
    async def enhance_property_image(self, image_data: str) -> str:
        """
        Enhance property images using AI
        """
        try:
            # Decode base64 image
            image_bytes = base64.b64decode(image_data)
            image = Image.open(io.BytesIO(image_bytes))
            
            # TODO: Apply AI enhancement
            # - Noise reduction
            # - Color correction
            # - Sharpening
            # - Lighting adjustment
            
            enhanced_image = await self._apply_enhancements(image)
            
            # Convert back to base64
            buffer = io.BytesIO()
            enhanced_image.save(buffer, format='PNG')
            img_str = base64.b64encode(buffer.getvalue()).decode()
            
            return img_str
            
        except Exception as e:
            raise Exception(f"Image enhancement failed: {str(e)}")
    
    async def analyze_property_features(self, image_data: str) -> Dict:
        """
        Analyze property features from images
        """
        try:
            # Decode base64 image
            image_bytes = base64.b64decode(image_data)
            image = Image.open(io.BytesIO(image_bytes))
            
            # TODO: Use AI to analyze property features
            # - Room detection
            # - Material identification
            # - Lighting analysis
            # - Space utilization
            
            analysis = await self._analyze_image_features(image)
            
            return analysis
            
        except Exception as e:
            raise Exception(f"Property analysis failed: {str(e)}")
    
    async def _create_2d_placeholder(
        self, 
        property_data: Dict, 
        style: str, 
        resolution: Tuple[int, int]
    ) -> Image.Image:
        """Create a sophisticated 2D placeholder"""
        width, height = resolution
        image = Image.new('RGB', (width, height), color='#f0f8ff')
        draw = ImageDraw.Draw(image)
        
        # Create a modern house layout
        house_width = width // 2
        house_height = height // 2
        house_x = (width - house_width) // 2
        house_y = (height - house_height) // 2
        
        # House outline
        draw.rectangle([house_x, house_y, house_x + house_width, house_y + house_height], 
                      fill='#ffffff', outline='#333333', width=3)
        
        # Roof
        roof_points = [
            (house_x, house_y),
            (house_x + house_width // 2, house_y - 50),
            (house_x + house_width, house_y)
        ]
        draw.polygon(roof_points, fill='#8b4513', outline='#654321', width=2)
        
        # Windows
        window_size = 40
        for i in range(2):
            for j in range(2):
                window_x = house_x + 50 + i * (house_width - 100 - window_size)
                window_y = house_y + 50 + j * (house_height - 100 - window_size)
                draw.rectangle([window_x, window_y, window_x + window_size, window_y + window_size],
                              fill='#87ceeb', outline='#333333', width=2)
        
        # Door
        door_width = 30
        door_height = 60
        door_x = house_x + house_width // 2 - door_width // 2
        door_y = house_y + house_height - door_height
        draw.rectangle([door_x, door_y, door_x + door_width, door_y + door_height],
                      fill='#8b4513', outline='#654321', width=2)
        
        # Add style-specific elements
        if style == "modern":
            # Modern elements
            draw.rectangle([house_x - 20, house_y + house_height - 20, 
                           house_x + house_width + 20, house_y + house_height + 20],
                          fill='#2c3e50', outline='#34495e', width=2)
        elif style == "traditional":
            # Traditional elements
            draw.ellipse([house_x + house_width - 30, house_y + 20, 
                         house_x + house_width - 10, house_y + 40],
                        fill='#ff6b6b', outline='#e74c3c', width=2)
        
        return image
    
    async def _create_3d_placeholder(
        self, 
        property_data: Dict, 
        style: str, 
        resolution: Tuple[int, int]
    ) -> Image.Image:
        """Create a 3D-style placeholder"""
        width, height = resolution
        image = Image.new('RGB', (width, height), color='#87ceeb')
        draw = ImageDraw.Draw(image)
        
        # Create 3D perspective house
        # Base rectangle (isometric view)
        base_points = [
            (width // 4, height // 2),
            (3 * width // 4, height // 2),
            (3 * width // 4, 3 * height // 4),
            (width // 4, 3 * height // 4)
        ]
        draw.polygon(base_points, fill='#ffffff', outline='#333333', width=3)
        
        # Front face
        front_points = [
            (width // 4, height // 2),
            (width // 2, height // 4),
            (3 * width // 4, height // 2),
            (3 * width // 4, 3 * height // 4),
            (width // 4, 3 * height // 4)
        ]
        draw.polygon(front_points, fill='#f8f9fa', outline='#333333', width=2)
        
        # Side face
        side_points = [
            (3 * width // 4, height // 2),
            (3 * width // 4, 3 * height // 4),
            (5 * width // 6, 2 * height // 3),
            (5 * width // 6, height // 3)
        ]
        draw.polygon(side_points, fill='#e9ecef', outline='#333333', width=2)
        
        # Roof (3D)
        roof_points = [
            (width // 2, height // 4),
            (width // 4, height // 2),
            (3 * width // 4, height // 2),
            (5 * width // 6, height // 3)
        ]
        draw.polygon(roof_points, fill='#8b4513', outline='#654321', width=2)
        
        return image
    
    async def _apply_enhancements(self, image: Image.Image) -> Image.Image:
        """Apply image enhancements"""
        # Convert to numpy array for processing
        img_array = np.array(image)
        
        # Apply basic enhancements
        # 1. Noise reduction
        img_array = cv2.bilateralFilter(img_array, 9, 75, 75)
        
        # 2. Contrast enhancement
        lab = cv2.cvtColor(img_array, cv2.COLOR_RGB2LAB)
        l, a, b = cv2.split(lab)
        clahe = cv2.createCLAHE(clipLimit=2.0, tileGridSize=(8,8))
        l = clahe.apply(l)
        img_array = cv2.merge([l, a, b])
        img_array = cv2.cvtColor(img_array, cv2.COLOR_LAB2RGB)
        
        # 3. Sharpening
        kernel = np.array([[-1,-1,-1], [-1,9,-1], [-1,-1,-1]])
        img_array = cv2.filter2D(img_array, -1, kernel)
        
        # Convert back to PIL Image
        enhanced_image = Image.fromarray(np.uint8(img_array))
        
        return enhanced_image
    
    async def _analyze_image_features(self, image: Image.Image) -> Dict:
        """Analyze property features in image"""
        # TODO: Implement actual AI analysis
        # For now, return sample analysis
        
        analysis = {
            "rooms_detected": [
                {"type": "living_room", "confidence": 0.95, "area": "large"},
                {"type": "kitchen", "confidence": 0.88, "area": "medium"},
                {"type": "bedroom", "confidence": 0.92, "area": "medium"}
            ],
            "materials": [
                {"type": "hardwood", "confidence": 0.85, "location": "flooring"},
                {"type": "marble", "confidence": 0.78, "location": "countertops"},
                {"type": "stainless_steel", "confidence": 0.91, "location": "appliances"}
            ],
            "lighting": {
                "natural_light": "excellent",
                "artificial_lighting": "modern",
                "overall_brightness": 0.85
            },
            "space_utilization": {
                "efficiency_score": 0.78,
                "open_concept": True,
                "storage_space": "adequate"
            },
            "style_analysis": {
                "primary_style": "modern",
                "color_scheme": "neutral",
                "design_quality": "high"
            }
        }
        
        return analysis

# Global AI engine instance
ai_engine = AIEngine()
