real-estate-visualizer/backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data processing and validation
pydantic==2.5.0
pydantic[email]==2.5.0
numpy==1.24.3
pandas==2.0.3

# Image processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# AI/ML dependencies (optional - for future AI model integration)
# torch==2.1.0
# transformers==4.35.2
# diffusers==0.24.0
